Create mockup of the WorkerManager.js
